==> 04-Cell-level_DGE.ipynb <==
# **Differential Gene Expression (Cell-level DGE)**

# **a. Same cell type across different conditions**

# Subset to one cell type
cell_type = "T cells"
adata_tcell = adata[adata.obs["celltypist_label"] == cell_type].copy()

adata_tcell

adata_tcell.obs['celltypist_label']

# Run DGE (tumor vs normal)
sc.tl.rank_genes_groups(
    adata_tcell,
    groupby="condition",
    groups=["PDAC"],
    reference="Normal",
    method="wilcoxon",
    pts=True
)

# View results
sc.pl.rank_genes_groups(
    adata_tcell,
    n_genes=25,
    sharey=False
)

# Export as DataFrame
dge_tcell_tumor_vs_normal = sc.get.rank_genes_groups_df(
    adata_tcell,
    group="PDAC"
)

dge_tcell_tumor_vs_normal.head()

# Export as csv
dge_tcell_tumor_vs_normal.to_csv(
    "dge_T-cells_pdac_vs_normal.csv",
    index=False
)

# **b. One cell type vs all other cell types in the same condition**

# Subset to one condition
adata_pdac = adata[adata.obs["condition"] == "PDAC"].copy()

adata_pdac.obs['condition']

# Create binary group (T cell vs Others)
adata_pdac.obs["tcell_vs_rest"] = np.where(
    adata_pdac.obs["celltypist_label"] == "T cells",
    "T cells",
    "rest"
)

# Run DGE
sc.tl.rank_genes_groups(
    adata_pdac,
    groupby="tcell_vs_rest",
    groups=["T cells"],
    reference="rest",
    method="wilcoxon",
    pts=True
)

# Plot
sc.pl.rank_genes_groups(
    adata_pdac,
    n_genes=25,
    sharey=False
)

# Extract results
dge_tcell_vs_others_pdac = sc.get.rank_genes_groups_df(
    adata_pdac,
    group="T cells"
)

dge_tcell_vs_others_pdac.head()

# Export as csv
dge_tcell_vs_others_pdac.to_csv(
    "dge_T-cells_vs_other_pdac.csv",
    index=False
)

# **c. One cell type vs all others regardless of condition**

# Create global binary label
adata.obs["tcell_global"] = np.where(
    adata.obs["celltypist_label"] == "T cells",
    "T cells",
    "rest"
)

# Run DGE
sc.tl.rank_genes_groups(
    adata,
    groupby="tcell_global",
    groups=["T cells"],
    reference="rest",
    method="wilcoxon",
    pts=True
)

# Plot
sc.pl.rank_genes_groups(
    adata,
    n_genes=25,
    sharey=False
)

# Extract results
dge_tcell_global = sc.get.rank_genes_groups_df(
    adata,
    group="T cells"
)

dge_tcell_global.head()

# Export as csv
dge_tcell_global.to_csv(
    "dge_T-cells_vs_other_no-condition.csv",
    index=False
)
